🚚 move: EvalLogList 관련 내용 EvalLogSearch -> shared
