+++ app/src/@shared/components/EvalLogList/EvalLogList.tsx
@@ -1,6 +1,6 @@
 import { EvalLog, TeamEvalLog } from '@shared/__generated__/graphql';
 import { VStack } from '@shared/ui-kit';
-import { EvalLogItem } from './EvalLogItem';
+import { EvalLogListItem } from './EvalLogListItem';
 
 type EvalLogListProps = {
   list: (EvalLog | TeamEvalLog)[];
@@ -10,7 +10,7 @@ export const EvalLogList = ({ list }: EvalLogListProps) => {
   return (
     <VStack as="ul" w="100%" spacing="1.6rem">
       {list.map((item) => (
-        <EvalLogItem key={item.id} element={item} />
+        <EvalLogListItem key={item.id} element={item} />
       ))}
     </VStack>
   );
